consolidate constants
