+++ napari/_qt/layers/qt_points_layer.py
@@ -9,7 +9,7 @@ from qtpy.QtWidgets import (
     QSlider,
 )
 
-from ...layers.points._points_constants import Mode, Symbol
+from ...utils.constants import PointsMode, Symbol
 from ..qt_color_dialog import QColorSwatchEdit
 from ..qt_mode_buttons import QtModePushButton, QtModeRadioButton
 from ..utils import qt_signals_blocked
@@ -118,13 +118,17 @@ class QtPointsControls(QtLayerControls):
         self.ndimCheckBox = ndim_cb
 
         self.select_button = QtModeRadioButton(
-            layer, 'select_points', Mode.SELECT, tooltip='Select points'
+            layer, 'select_points', PointsMode.SELECT, tooltip='Select points'
         )
         self.addition_button = QtModeRadioButton(
-            layer, 'add_points', Mode.ADD, tooltip='Add points'
+            layer, 'add_points', PointsMode.ADD, tooltip='Add points'
         )
         self.panzoom_button = QtModeRadioButton(
-            layer, 'pan_zoom', Mode.PAN_ZOOM, tooltip='Pan/zoom', checked=True
+            layer,
+            'pan_zoom',
+            PointsMode.PAN_ZOOM,
+            tooltip='Pan/zoom',
+            checked=True,
         )
         self.delete_button = QtModePushButton(
             layer,
@@ -207,11 +211,11 @@ class QtPointsControls(QtLayerControls):
             Raise error if event.mode is not ADD, PAN_ZOOM, or SELECT.
         """
         mode = event.mode
-        if mode == Mode.ADD:
+        if mode == PointsMode.ADD:
             self.addition_button.setChecked(True)
-        elif mode == Mode.SELECT:
+        elif mode == PointsMode.SELECT:
             self.select_button.setChecked(True)
-        elif mode == Mode.PAN_ZOOM:
+        elif mode == PointsMode.PAN_ZOOM:
             self.panzoom_button.setChecked(True)
         else:
             raise ValueError("Mode not recognized")
