--- napari/components/dims_constants.py
@@ -1,6 +1,0 @@
-from enum import Enum
-
-
-class DimsMode(Enum):
-    POINT = 0
-    INTERVAL = 1
