+++ napari/utils/misc.py
@@ -5,7 +5,6 @@ import inspect
 import itertools
 import re
 
-from enum import Enum, EnumMeta
 from os import PathLike, fspath, path
 from typing import Optional, Sequence, Type, TypeVar
 
@@ -119,66 +118,6 @@ def formatdoc(obj):
         return obj
     finally:
         del frame
-
-
-class StringEnumMeta(EnumMeta):
-    def __getitem__(self, item):
-        """ set the item name case to uppercase for name lookup
-        """
-        if isinstance(item, str):
-            item = item.upper()
-
-        return super().__getitem__(item)
-
-    def __call__(
-        cls,
-        value,
-        names=None,
-        *,
-        module=None,
-        qualname=None,
-        type=None,
-        start=1,
-    ):
-        """ set the item value case to lowercase for value lookup
-        """
-        # simple value lookup
-        if names is None:
-            if isinstance(value, str):
-                return super().__call__(value.lower())
-            elif isinstance(value, cls):
-                return value
-            else:
-                raise ValueError(
-                    f'{cls} may only be called with a `str`'
-                    f' or an instance of {cls}'
-                )
-
-        # otherwise create new Enum class
-        return cls._create_(
-            value,
-            names,
-            module=module,
-            qualname=qualname,
-            type=type,
-            start=start,
-        )
-
-    def keys(self):
-        return list(map(str, self))
-
-
-class StringEnum(Enum, metaclass=StringEnumMeta):
-    def _generate_next_value_(name, start, count, last_values):
-        """ autonaming function assigns each value its own name as a value
-        """
-        return name.lower()
-
-    def __str__(self):
-        """String representation: The string method returns the lowercase
-        string of the Enum name
-        """
-        return self.value
 
 
 camel_to_snake_pattern = re.compile(r'(.)([A-Z][a-z]+)')
